Convert from Gradle compatibility to toolchain
We used to use the `sourceCompatibility` and `targetCompatibility` variables to indicate which version of Java we wanted Gradle to use. This switches to the new toolchain option, which will actually download the specified JDK from AdoptOpenJDK if it's not already available on the system.

That's a bit slow the first time you run Gradle (since it's installing a full JDK), but is fast enough after that, and it may simplify some of the occasional issues of JDK mismatches.

Co-authored-by: Nic McPhee <mcphee@morris.umn.edu>