Rename Database to UserDatabase
In going over the lab in class today, I stumbled across an issue that, while minor, has come up repeatedly over the years. We (& I think it's "I", TBH) called the "fake database" class `Database`. This creates an awkward moment when the students need to make their own to-do database class.

I think that renaming `Database` to `UserDatabase` would help with that in the future. I ended up showing them how to do that in VS Code (so nice), so I figured I'd turn that into a PR. If people are OK with it, it will hopefully make things at least a _little_ better in the future.