Address most of the Checkstyle issues
This deals with most of the Checkstyle issues with our codebase. There were a bunch of magic numbers in the tests, so I again just ignored them. (Not thrilled about that, but I'm not sure what the better plan would be short of a complete rewrite.)

There were also a number of places where I had spaces after open curly braces in the initialization of arrays, and Checkstyle didn't like those. I don't care much, though, and am happy to remove them.

There are still a number of complaints about the naming of our spec file functions. I'll deal with those in a separate commit, though.