Fix failing test in `UserControllerSpec`
When I changed things in the previous commit, I deleted the line that set the status to 201 when we successfully found a user. @helloworld12321 had pointed out that this wasn't the correct code, and I somehow thought I didn't need to set an explicit code in this case.

Turns out I *do* need to set an explicit code if we're going to check for it in the tests (which we do).

I also discovered, though, that there's an `HttpCode` class in Javalin that does a nice job of including all the standard HTTP response codes, so I switched to using that.