Add Javalin functional tests
This follows the example from "Testing Javalin Applications" (https://javalin.io/tutorials/testing) for adding functional tests to a Javalin app. This does give us very nearly 100% test coverage, including coverage of the `Server` class, which was entirely uncovered before.

The tests in `ServerFunctionalSpec` are a little odd, mostly because the examples that they test are unusual (or, in the case of "/hello", totally unnecessary).

One could essentially replace all/most of the unit tests with functional tests like these, and still get quite reasonable coverage. That wouldn't be the recommended approach, but it would probably be where students would gravitate since it's often easier to think about "big picture" E2E-style tests.