Use encodeURIComponent to encode field values
As @helloworld12321 pointed out in #170, it's possible for users to "inject" data into the query params via the fields in our simple user interface.

We should be demonstrated reasonable (if we don't always manage *best* :->) practices, so it seemed good to encode whatever people enter in a field in this lab.

There are probably similar things that should be done in later labs as well.

Closes #170