Revert the removal of the redirects
For now I want to keep the redirects. I realize that they don't really do much, but I think keeping them as an example of something that Javalin can do seems like it might be useful.

I retained the addition of leading slashes in the `link` and `script` tags in the HTML since that seemed to be a separate issue.