Address Checkstyle issues
I forgot to run Checkstyle on the last commit, and I'd introduced a few problems that caused the build to fail on GitHub.

Most of this was just long lines, but there is a fight going on between VS Code and Checkstyle about whether there should be spaces after `{` in initializing arrays. I actually like the space, so perhaps we want to disable or tweak the Checkstyle check.