Remove JSON translation from functional tests
@floogulinc pointed out that we didn't really need to have all the JSON translation in the functional tests, and we could just compare the POJOs directly. This changes the tests to work on POJOs.

To make this work I had to implement an `equals()` method for the `User` class; otherwise testing for equality failed because the object we got from the DB was different than the object we got from the response. To keep things simple I implemented this to just test to see if the `_id`s were the same, but we could make that fancier and check the other fields as well.

Implementing `equals()` meant that I had to implement `hashCode()`, which just returns the hash of the ID string.

I also added a simple `toString()` that prints the name to aid in some debugging.