Convert server config to Javalin v5
In going from v5 to v5, Javalin changed the configuration system a bit, so we now have to get the objects holding the static files/plugins, and then add static files/plugins to those objects. Luckily the argument structure was all the same, so this was pretty easy to sort out.