Remove clearing the cookie store on the mock context
I'm honestly not sure why the line `ctx.clearCookieStore()` was added to the `setUp()` code in the `UserControllerSpec` tests. It doesn't make a lot of sense to clear the cookie store on a mock object, and I'm not entirely sure how that ever worked or did anything?

Anyway, getting the cookie store and clearing it (the v5 Javalin way) throws a `NullPointerException`, presumably because `ctx.cookieStore()` returns `null`. So I just removed the line and everything seems to work.