Removes test that gets users of a certain age
There is a test that models this behavior better that already exists in UserControllerSpec.java. This commit also adds an assertion to the test of filtering by age to see that the right number of users were returned in that filter for age 25. If others consider this assertion too fragile, it would be fine to remove this assertion.