Removes test that gets user by ID
This commit removes a test that gets user by ID since there is a better model of such a test in UserControllerSpec.java. This commit also adds another assertion to the test of getting user by ID. If this assertion is too fragile, please comment on this at the time of the PR and let's not include it.