Removes test that gets users with two filters
Removes the test of getting users who work for OHMNET and are age 25 (tested for each separately and then with those filters combined). Adds an assertion to the test in UserControllerSpec.java that checks that there is just one such user. If this assertion is too fragile, please comment to suggest that change in the PR.